+++ setup.cfg
@@ -90,8 +90,8 @@ console_scripts =
 # CAUTION: --cov flags may prohibit setting breakpoints while debugging.
 #          Comment those flags to avoid this pytest issue.
 addopts =
-    --cov sxapi --cov-report term-missing
-    --verbose
+    #--cov sxapi --cov-report term-missing
+    #--verbose
 norecursedirs =
     dist
     build
@@ -120,6 +120,8 @@ exclude =
     dist
     .eggs
     docs/conf.py
+per-file-ignores =
+    */__init__.py: F401
 
 [isort]
 # configurations for isort import style checker
