+++ src/sxapi/cli/example-config.conf
@@ -4,7 +4,7 @@
 USER = user.name@example.com
 
 # smaXtec password
-PASSWORD = "userSecret"
+PASSWORD = userSecret
 
 # organisation to retrieve data from
 ORGA = smaxtec_organisation_id
