+++ src/sxapi/cli/parser/subparser/__init__.py
@@ -1,0 +1,1 @@
+from sxapi.cli.parser.subparser.token import SxApiTokenSubparser
