+++ tests/cli_tests/test-config.conf
@@ -4,5 +4,5 @@ USER = test@example.com
 PASSWORD = smaxtec_test_user_pwd
 ORGA = smaxtec_test_organisation_id
 
-API_PUBLIC_V2_PATH = test_path_test
+API_PUBLIC_V2_PATH = https://test_path_test/users/credentials
 API_INTEGRATION_V2_PATH = test_path_integration
